requirements.txt
aiosqlite==0.19.0
camelot-py==0.11.0
fastapi==0.95.1
fastapi-users[sqlalchemy]==11.0.0
ghostscript==0.7
google-cloud-documentai==2.15.0
gradio==3.27.0
langchain==0.0.188
llama-index==0.6.15
openai==0.27.3
opencv-python==4.7.0.72
pandas==1.5.3
passlib==1.7.4
pdfplumber==0.8.0
pydantic==1.10.7
python-jose==3.3.0
starlette==0.26.1
tenacity==8.2.2
tiktoken==0.3.2
uvicorn==0.22.0
